requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
